pyproject.toml
[project]
name = "viajante"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
    "schedule>=1.2.2",
]
